+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.77.2
+
+### Command-Line Interface
+
+* Properly handle the `--silence-deprecation` flag.
+
+* Handle the `--fatal-deprecation` and `--future-deprecation` flags for
+  `--interactive` mode.
+
 ## 1.77.1
 
 * Fix a crash that could come up with importers in certain contexts.